+++ apps/docs/src/examples/extras/image-material/BentPlaneGeometry.svelte
