+++ apps/docs/src/examples/extras/image-material/rgbaProcessingTexture/fragmentShader.glsl
