+++ apps/docs/src/examples/extras/image-material/rgbaProcessingTexture/vertexShader.glsl
