+++ packages/extras/src/lib/components/ImageMaterial/ImageMaterial.svelte
@@ -1,20 +1,26 @@
 <script lang="ts">
-  import { Color, Vector2, Mesh, type Texture, Vector3 } from 'three'
-  import { T, asyncWritable, forwardEventHandlers, useThrelte } from '@threlte/core'
+  import { Color, Vector2, type Texture, type Mesh, Vector3, ShaderMaterial } from 'three'
+  import {
+    T,
+    asyncWritable,
+    forwardEventHandlers,
+    useParent,
+    useTask,
+    useThrelte
+  } from '@threlte/core'
   import { useTexture } from '../../hooks/useTexture'
-  import type { ImageProps, ImageEvents, ImageSlots } from './Image'
+  import type { ImageMaterialProps, ImageMaterialEvents, ImageMaterialSlots } from './ImageMaterial'
   import { useSuspense } from '../../suspense/useSuspense'
   import { vertexShader, fragmentShader } from './shaders'
 
-  type $$Props = ImageProps
-  type $$Events = ImageEvents
-  type $$Slots = ImageSlots
+  type $$Props = ImageMaterialProps
+  type $$Events = ImageMaterialEvents
+  type $$Slots = ImageMaterialSlots
 
-  type Props = Required<ImageProps>
+  type Props = Required<ImageMaterialProps>
 
   export let color: Props['color'] = 'white'
-  export let segments: Props['segments'] = 1
-  export let scale: Props['scale'] = 1
+  export let scale: ImageMaterialProps['scale'] = undefined
   export let zoom: Props['zoom'] = 1
   export let radius: Props['radius'] = 0
   export let alphaProgress: Props['alphaProgress'] = 0
@@ -24,29 +30,28 @@
   export let hue: Props['hue'] = 0
   export let saturation: Props['saturation'] = 0
   export let lightness: Props['lightness'] = 0
-  export let monochromeColor: ImageProps['monochromeColor'] = undefined
-  export let monochromeStrength: ImageProps['monochromeStrength'] = undefined
+  export let monochromeColor: ImageMaterialProps['monochromeColor'] = undefined
+  export let monochromeStrength: ImageMaterialProps['monochromeStrength'] = undefined
   export let negative: Props['negative'] = false
-  export let colorProcessingTexture: ImageProps['colorProcessingTexture'] = undefined
+  export let colorProcessingTexture: ImageMaterialProps['colorProcessingTexture'] = undefined
   export let opacity: Props['opacity'] = 1
-  export let texture: ImageProps['texture'] = undefined
+  export let texture: ImageMaterialProps['texture'] = undefined
   export let toneMapped: Props['toneMapped'] = true
   export let transparent: Props['transparent'] = false
-  export let side: ImageProps['side'] = undefined
-  export let url: ImageProps['url'] = undefined
+  export let side: ImageMaterialProps['side'] = undefined
+  export let url: ImageMaterialProps['url'] = undefined
 
-  export let ref = new Mesh()
+  export let ref = new ShaderMaterial()
 
   const suspend = useSuspense()
-  $: textureStore = url ? suspend(useTexture(url)) : asyncWritable(Promise.resolve(texture))
+  $: textureStore = suspend(url ? useTexture(url) : asyncWritable(Promise.resolve(texture)))
 
   let { size } = useThrelte()
+  const parent = useParent()
 
   const uniforms = {
     color: { value: new Color(color) },
-    scale: {
-      value: typeof scale === 'number' ? new Vector2(scale, scale) : new Vector2(scale[0], scale[1])
-    },
+    scale: { value: new Vector2() },
     imageBounds: { value: new Vector2(1, 1) },
     resolution: { value: 1024 },
     map: { value: null as Texture | null },
@@ -67,12 +72,6 @@
   }
 
   $: uniforms.color.value.set(color)
-
-  $: if (typeof scale === 'number') {
-    uniforms.scale.value.set(scale, scale)
-  } else {
-    uniforms.scale.value.set(scale[0], scale[1])
-  }
 
   $: uniforms.imageBounds.value.set(
     $textureStore?.image.width ?? 0,
@@ -126,31 +125,52 @@
     uniforms.colorProcessingEnabled.value = colorProcessingEnabled
   }
 
-  // Support arbitrary plane geometries (for instance with rounded corners)
-  $: if ('parameters' in ref.geometry) {
-    const { width, height } = ref.geometry.parameters as { width: number; height: number }
-    uniforms.scale.value.set(uniforms.scale.value.x * width, uniforms.scale.value.y * height)
+  $: if (typeof scale === 'number') {
+    uniforms.scale.value.set(scale, scale)
+  } else if (Array.isArray(scale)) {
+    uniforms.scale.value.set(scale[0], scale[1])
+  }
+
+  const { start, stop } = useTask(
+    () => {
+      const mesh = $parent
+
+      if (mesh === undefined || !('isMesh' in mesh)) {
+        return
+      }
+
+      uniforms.scale.value.set(mesh.scale.x, mesh.scale.y)
+
+      const geometry = (mesh as Mesh).geometry
+
+      // Support arbitrary plane geometries (for instance with rounded corners)
+      if ('parameters' in geometry) {
+        const { width, height } = geometry.parameters as { width: number; height: number }
+        uniforms.scale.value.set(uniforms.scale.value.x * width, uniforms.scale.value.y * height)
+      }
+    },
+    { autoStart: false }
+  )
+
+  $: if (scale === undefined) {
+    start()
+  } else {
+    stop()
   }
 
   const component = forwardEventHandlers()
 </script>
 
-{#if $textureStore}
-  <T
-    is={ref}
-    scale={typeof scale === 'number' ? [scale, scale, 1] : [scale[0], scale[1], 1]}
-    {...$$restProps}
-    bind:this={$component}
-  >
-    <T.PlaneGeometry args={[1, 1, segments, segments]} />
-    <T.ShaderMaterial
-      {uniforms}
-      {toneMapped}
-      {transparent}
-      {side}
-      {vertexShader}
-      {fragmentShader}
-    />
-    <slot {ref} />
-  </T>
-{/if}
+<T
+  is={ref}
+  {uniforms}
+  {toneMapped}
+  {transparent}
+  {side}
+  {vertexShader}
+  {fragmentShader}
+  {...$$restProps}
+  bind:this={$component}
+>
+  <slot {ref} />
+</T>
