+++ packages/extras/src/lib/components/ImageMaterial/shaders.ts
