+++ packages/extras/src/lib/index.ts
@@ -17,7 +17,7 @@ export { default as GLTF } from './components/GLTF/GLTF.svelte'
 export { default as Gizmo } from './components/Gizmo/Gizmo.svelte'
 export { default as ContactShadows } from './components/ContactShadows/ContactShadows.svelte'
 export { default as Environment } from './components/Environment/Environment.svelte'
-export { default as Image } from './components/Image/Image.svelte'
+export { default as ImageMaterial } from './components/ImageMaterial/ImageMaterial.svelte'
 export { default as Grid } from './components/Grid/Grid.svelte'
 export { default as RoundedBoxGeometry } from './components/RoundedBoxGeometry/RoundedBoxGeometry.svelte'
 export { default as TransformControls } from './components/controls/TransformControls/TransformControls.svelte'
